refactor: world to string function
The method that creates printable strings from the world is now part of
the context - that way it can directly use the configured colors. The
function now works with a two dimensional vector of strings that are
filled by the objects in the world. Only in the end this two dimensional
vector is then converted into lines.